% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlopeFinder.R
\name{findSlope.beta.threePoints}
\alias{findSlope.beta.threePoints}
\title{Slope Finding using Beta Method (3 points)}
\usage{
findSlope.beta.threePoints(
  data,
  x = point.x(data),
  y = point.y(data),
  betaSolver = solve.beta
)
}
\arguments{
\item{data}{A \code{pointData} type that stores the 3 points. It will be converted into \code{x} and \code{y}.}

\item{x}{A vector of x-coordinates of the 3 points. May be assigned directly.}

\item{y}{A vector of y-coordinates of the 3 points. May be assigned directly.}

\item{betaSolver}{A function to solve beta in the form of \code{function(A, B, b)} that returns a list containing beta. Can be \code{solve.beta} (default) or \code{solve.beta.cholesky}.}
}
\value{
The slopes at the points of interest.
}
\description{
Slope Finding using Beta Method (3 points)
}
