% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpolation.R
\name{interpolate.patch.onePointSlope}
\alias{interpolate.patch.onePointSlope}
\title{Interpolation by Patching Functions Generated at Each Point}
\usage{
interpolate.patch.onePointSlope(
  data,
  slopes,
  solver,
  patch = getOption("defaultPatching", default = patch.fifthDegree)
)
}
\arguments{
\item{data}{A \code{pointData} type that stores all the points to be interpolated.}

\item{slopes}{A \code{vector} type that stores all the slopes correspond to each point.}

\item{solver}{Function that returns a polynomial that is generated at a point with slope. Must be in the form of \verb{function(data, slope)}.}

\item{patch}{The function used for patching, uses \code{defaultPatchPolynomial} by default. Must be in the form of \verb{function(a,b,p)}.}
}
\value{
A piecewise polynomial.
}
\description{
Interpolation by Patching Functions Generated at Each Point
}
