% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpolation.R
\name{interpolate.patch.threePoint}
\alias{interpolate.patch.threePoint}
\title{Interpolation by Patching Three Points Segments}
\usage{
interpolate.patch.threePoint(data, solver, patch = patch.fifthDegree)
}
\arguments{
\item{data}{A \code{pointData} type that stores all the points to be interpolated.}

\item{solver}{Function that returns a polynomial that interpolate 3 points, must be in the form of \verb{function(data)}.}

\item{patch}{The function used for patching, uses \code{defaultPatchPolynomial} by default. Must be in the form of \verb{function(a,b,p)}.}
}
\value{
A piecewise polynomial.
}
\description{
Interpolation by Patching Three Points Segments
}
