% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpolation.R
\name{linearPolynomial}
\alias{linearPolynomial}
\title{Linear Polynomial Given Point and Slope}
\usage{
linearPolynomial(data, x = point.x(data), y = point.y(data), slope)
}
\arguments{
\item{x}{The vector of x-coordinates. May be assigned directly.}

\item{y}{The vector of y-coordinates. May be assigned directly.}

\item{slope}{The slope at the point.}
}
\value{
The linear polynomial that goes through the point with the given slope.
}
\description{
Linear Polynomial Given Point and Slope
}
