% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuadraticProgramming.R
\name{min.quadratic}
\alias{min.quadratic}
\title{Minimize Quadratic with Linear Constraints

Minimizes \verb{x^T A x + x^T B} under equality constraint \verb{E x = F} and inequality constraint \verb{G x >= H}}
\usage{
\method{min}{quadratic}(
  A,
  B,
  E = NULL,
  F = NULL,
  G = NULL,
  H = NULL,
  tol = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{A}{A symmetric matrix describing the quadratic coefficients}

\item{B}{A vector describeing the linear coefficients}

\item{E}{A matrix for equality constraint}

\item{F}{A vector for equality constraint}

\item{G}{A matrix for inequality constraint}

\item{H}{A vector for inequality constraint}

\item{tol}{Tolerance. Number whose absolute value is less than \code{tol} is considered 0.}
}
\value{
A list containing:
\code{x}: the desired \code{x}, see main description
\code{value}: the minimum value of \verb{x^T A x + x^T B}
\code{lambda}: (optional) the Lagrangian multipliers for the equality constraints
\code{lambda}: (optional) the Lagrangian multipliers for the inequality constraints
}
\description{
Minimize Quadratic with Linear Constraints

Minimizes \verb{x^T A x + x^T B} under equality constraint \verb{E x = F} and inequality constraint \verb{G x >= H}
}
\details{
Uses Karush-Kuhn-Tucker method internally.
}
\examples{
A <- matrix(c(1, 0, 0, 2), nrow = 2, ncol = 2)
B <- c(-4, -4)
G <-matrix(c(-1, 1, -4, -1), nrow = 2, ncol = 2)
H <- c(-3, 0)
result <- min.quadratic(A, B, G = G, H = H)
}
