% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patching.R
\docType{class}
\name{patchingFunction-class}
\alias{patchingFunction-class}
\alias{patchingFunction}
\title{An S4 Class to Represent Patching Function}
\description{
An S4 Class to Represent Patching Function
}
\details{
A well behaved patching function \code{theta} should evaluate to 1 at 0, and 0 at 1.
}
\section{Slots}{

\describe{
\item{\code{theta}}{A function in the format of \verb{function(x)} that returns a numeric value.}

\item{\code{description}}{A function in the format of \verb{function(a, b)} where \code{a} and \code{b} are left and right bounds of an interval and returns a string.}
}}

