% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patching.R
\docType{class}
\name{patchingPolynomial-class}
\alias{patchingPolynomial-class}
\alias{patchingPolynomial}
\title{Polynomial Patching Function}
\description{
Polynomial Patching Function
}
\details{
It inherits from \code{patchingFunction} class.
}
\section{Slots}{

\describe{
\item{\code{polynomial}}{A \code{polynomial} type.}
}}

