% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpolation.R
\name{percentagePolynomial}
\alias{percentagePolynomial}
\title{Percentage Polynomial

The linear function that is 0 when evaluated at \code{min} and 1 when evaluated at \code{max}}
\usage{
percentagePolynomial(min, max)
}
\arguments{
\item{min}{The value at which the function is 0.}

\item{max}{The value at which the function is 1.}
}
\value{
\code{(x-min)/(max-min)}
}
\description{
Percentage Polynomial

The linear function that is 0 when evaluated at \code{min} and 1 when evaluated at \code{max}
}
