% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PiecewisePolynomial.R
\docType{class}
\name{piecewisePolynomial-class}
\alias{piecewisePolynomial-class}
\alias{piecewisePolynomial}
\title{An S4 Class to Represent a Piecewise Polynomial}
\description{
An S4 Class to Represent a Piecewise Polynomial
}
\details{
Use \code{length(object)} to get the number of piecewise ranges.
Use \code{predict(object, newdata)} to evaluate at \code{newdata}. Alternatively, use \code{as.function(x)} to turn the piecewise polynomial into a function.
Use \code{differentiate(x)} to take the derivative of each polynomial.
Use \code{plot(x)} to \code{lines{x}} to plot the piecewise polynomial.
}
\section{Slots}{

\describe{
\item{\code{leftBound}}{A vector that stores the left bounds of the piecewise ranges.}

\item{\code{rightBound}}{A vector that stores the right bounds of the piecewise ranges.}

\item{\code{polynomial}}{A list that stores the polynomials at each piecewise ranges.}
}}

