% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PiecewisePolynomial.R
\name{piecewisePolynomial.extrema}
\alias{piecewisePolynomial.extrema}
\title{Finding the Extrema of a Piecewise Polynomial}
\usage{
piecewisePolynomial.extrema(poly, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{poly}{A \code{piecewisePolynomial} type.}

\item{tol}{Tolerance.}
}
\value{
A \code{pointData} type containing the extrema points.
}
\description{
The ranges in this Piecewise Polynomial must be ordered.
The boundary points are also includede in the result.
}
