% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PiecewisePolynomial.R
\name{piecewisePolynomial.extrema.x}
\alias{piecewisePolynomial.extrema.x}
\title{Finding the x-values of Extrema of a Piecewise Polynomial}
\usage{
piecewisePolynomial.extrema.x(poly, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{poly}{A \code{piecewisePolynomial} type.}

\item{tol}{Tolerance.}
}
\value{
A vector containing the x-values of the extrema.
}
\description{
Finding the x-values of Extrema of a Piecewise Polynomial
}
