% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointData.R
\docType{class}
\name{pointData-class}
\alias{pointData-class}
\alias{pointData}
\title{An S4 Class to Represent a Points}
\description{
An S4 Class to Represent a Points
}
\details{
Use \code{length(object)} to get the number of points.
Use \code{point.x(object)} or \code{point.y(object)} to access the x-coordinates or y-coordinates, respectively.
Use \code{differentiate(x)} to take the derivative of each polynomial.
Use \code{points(x)} to plot the points.
Use \code{[]} to access a subset of points.
}
\section{Slots}{

\describe{
\item{\code{x}}{A vector that stores x-coordinates.}

\item{\code{y}}{A vector that stores y-coordinates.}
}}

