% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polynomial.R
\name{polynomial.extrema}
\alias{polynomial.extrema}
\title{Finding the Extrema of a Polynomial}
\usage{
polynomial.extrema(poly, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{poly}{A \code{polynomial} type.}

\item{tol}{Tolerance.}
}
\value{
A \code{pointData} type containing the extrema points.
}
\description{
Finding the Extrema of a Polynomial
}
