% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polynomial.R
\name{polynomial.extrema.x}
\alias{polynomial.extrema.x}
\title{Finding the x-values of Extrema of a Polynomial}
\usage{
polynomial.extrema.x(poly, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{poly}{A \code{polynomial} type.}

\item{tol}{Tolerance.}
}
\value{
A vector containing the x-values of the extrema.
}
\description{
Finding the x-values of Extrema of a Polynomial
}
