% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polynomial.R, R/PiecewisePolynomial.R
\name{polynomial.extrema.y}
\alias{polynomial.extrema.y}
\title{Finding the y-values of Extrema of a Polynomial}
\usage{
polynomial.extrema.y(poly, tol = sqrt(.Machine$double.eps))

polynomial.extrema.y(poly, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{poly}{A \code{piecewisePolynomial} type.}

\item{tol}{Tolerance.}
}
\value{
A vector containing the y-values of the extrema.

A vector containing the y-values of the extrema.
}
\description{
Finding the y-values of Extrema of a Polynomial

Finding the y-values of Extrema of a Piecewise Polynomial
}
