% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpolation.R
\name{quadratic.point.slope.extrema}
\alias{quadratic.point.slope.extrema}
\title{Quadratic Function Through a Point and with a Given Slope and with Given y-value of Extrema}
\usage{
quadratic.point.slope.extrema(
  data,
  x = point.x(data),
  y = point.y(data),
  slope,
  extrema = 0,
  tol = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{data}{A \code{pointData} type containing a signle point. It will be converted to \code{x} and \code{y}}

\item{x}{The x-coordinate of a point. May be assigned directly.}

\item{y}{The y-coordinate of a point. May be assigned directly.}

\item{slope}{The slope at point `(x,y)``}

\item{extreme}{The y-value of the extrema.}
}
\value{
The quadratic that goes through \verb{(x,y)} with slope \code{k} and tangent to the \code{y=extrema}.
}
\description{
Quadratic Function Through a Point and with a Given Slope and with Given y-value of Extrema
}
