% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpolation.R
\name{quadraticPolynomial}
\alias{quadraticPolynomial}
\title{Quadratic Function Through Three Points}
\usage{
quadraticPolynomial(data, x = point.x(data), y = point.y(data))
}
\arguments{
\item{x}{The vector of x-coordinates. May be assigned directly.}

\item{y}{The vector of y-coordinates. May be assigned directly.}
}
\value{
The quadratic polynomial that goes through \code{(x[1],y[1])}, \code{(x[2],y[2])}, and \code{(x[3],y[3])}.
}
\description{
Quadratic Function Through Three Points
}
