% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpolation.R
\name{rrinterpolate}
\alias{rrinterpolate}
\title{Interpolation with Boundaries}
\usage{
rrinterpolate(x, y, min = NA, max = NA)
}
\arguments{
\item{x}{A vector of x-values of the points to be interpolated over.}

\item{y}{A vector of y-values of the points to be interpolated over.}

\item{min}{Optional lower boundary of the interpolation.}

\item{max}{Optional upper boundary of the interpolation.}
}
\value{
A \code{patching} S4 class.
}
\description{
The algorithm produces a C^2 curve that interpolates the specified points while staying between \code{min} and \code{max} boundaries. The resulting function will have absolute second derivative within some constant multiplied by the optimal (the optimal being the theoretical minimal second derivative possible for a C^2 curve interpolating through the points).
}
