% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpolation.R
\name{rrinterpolate.slope}
\alias{rrinterpolate.slope}
\title{Interpolation with Boundaries Given Prescribed Slopes}
\usage{
rrinterpolate.slope(x, y, k, min, max)
}
\arguments{
\item{x}{A vector of x-values of the points to be interpolated over.}

\item{y}{A vector of y-values of the points to be interpolated over.}

\item{k}{A vector that describes the slopes at each point.}

\item{min}{Optional lower boundary of the interpolation.}

\item{max}{Optional upper boundary of the interpolation.}
}
\value{
A \code{patching} S4 class.
}
\description{
Interpolation with Boundaries Given Prescribed Slopes
}
\seealso{
\code{\link[=rrinterpolate]{rrinterpolate()}} for interpolation details.
}
