% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{rrplot}
\alias{rrplot}
\title{Plot Interpolation

Perform plotting of function of the \code{patching} S4 class. It will plot the function along with its first and second derivatives.}
\usage{
rrplot(interpolation, interval, x.point, y.point, limits, res = 1000, autodiff)
}
\arguments{
\item{interpolation}{A \code{patching} class.}

\item{interval}{A vector of x-values to be plotted.}

\item{x.point}{A vector of x-values of the points to be plotted.}

\item{y.point}{A vector of y-values of the points to be plotted.}

\item{limits}{A vector y-values of the horizontal lines to be plotted.}

\item{res}{The number of points are used to calculate the function when \code{interval} is not specified.}

\item{autodiff}{Whther to use automatic differentiation to calculate the derivatives. When the \code{patching} class can be collapsed into a polynomial, this parameter can be set to \code{FALSE} to improve algorithm efficiency. When this parameter is missing, it determines automatically whether the algorithm can be collapsed into apolynomial.}
}
\description{
Plot Interpolation

Perform plotting of function of the \code{patching} S4 class. It will plot the function along with its first and second derivatives.
}
\details{
Packages "reshape2" and "ggplot2" are required for this action.
}
