% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QuadraticProgramming.R
\name{solve.beta}
\alias{solve.beta}
\title{Solving for beta}
\usage{
\method{solve}{beta}(A, B, b, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{A}{A 6x6 symmetric postive-semidefinite matrix}

\item{B}{A 6x6 matrix}

\item{b}{A vector of 6 numbers}

\item{tol}{Tolerance. Number whose absolute value is less than \code{tol} is considered 0.}
}
\value{
the desired \code{beta}, see main description
}
\description{
Find \code{beta} such that \verb{t(beta) \%*\% A \%*\% beta + ||beta||_1} is minimzied subject to \code{B \%*\% beta = b}. \verb{||beta||_1} is the 1-norm of \code{beta}
}
\examples{
A <- diag(c(1 / 2, 1 / 3, 1 / 4, 1, 2, 3))
B <- diag(c(2, 3, 6, 0, 0, 0))
b <- c(1, 1, 1, 0, 0, 0)
result <- solve.beta(A, B, b)
}
